requirements.txt
os-sys
openai-whisper
pandas
soundfile
librosa
streamlit
spacy
nltk
matplotlib
